pyproject.toml
[project]
name = "Slack"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Edgar E. Reyes T.", email = "linuosx@gmail.com"},
]
dependencies = ["requests>=2.32.4", "python-decouple>=3.8"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
